# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gas Monitor API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # ── Devices ───────────────────────────────────────────────────────────
  /devices:
    get:
      operationId: getDevices
      tags: [devices]
      summary: List devices for current homeowner
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
    post:
      operationId: createDevice
      tags: [devices]
      summary: Register a new ESP32 device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{id}:
    get:
      operationId: getDevice
      tags: [devices]
      summary: Get device by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          description: Not found
    patch:
      operationId: updateDevice
      tags: [devices]
      summary: Update device
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
    delete:
      operationId: deleteDevice
      tags: [devices]
      summary: Delete device
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content

  # ── Sensor Readings ───────────────────────────────────────────────────
  /devices/{deviceId}/readings:
    get:
      operationId: getReadings
      tags: [readings]
      summary: Get last 24 sensor readings for a device
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SensorReading"
    post:
      operationId: createReading
      tags: [readings]
      summary: Submit a sensor reading (from ESP32)
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SensorReadingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SensorReading"
  /devices/{deviceId}/readings/latest:
    get:
      operationId: getLatestReading
      tags: [readings]
      summary: Get the most recent sensor reading
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SensorReading"
        "404":
          description: No readings yet

  # ── Alerts ────────────────────────────────────────────────────────────
  /alerts:
    get:
      operationId: getAlerts
      tags: [alerts]
      summary: Get alerts for current user
      parameters:
        - name: resolved
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{id}/resolve:
    patch:
      operationId: resolveAlert
      tags: [alerts]
      summary: Mark alert as resolved
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"

  # ── Refill Orders ─────────────────────────────────────────────────────
...
  /messages:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get conversation messages with a user
      parameters:
        - name: withUserId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
components:
  schemas:
...
    UserRegistration:
      type: object
      required: [name, email, password, role]
      properties:
        name:    { type: string }
        email:   { type: string }
        password: { type: string, minLength: 6 }
        role:    { type: string, enum: [homeowner, supplier] }
    UserLogin:
      type: object
      required: [email, password]
      properties:
        email:    { type: string }
        password: { type: string }
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:        { type: integer }
        name:      { type: string }
        email:     { type: string }
        role:      { type: string, enum: [homeowner, supplier] }
        createdAt: { type: string, format: date-time }
    AuthResponse:
      type: object
      required: [user]
      properties:
        user: { $ref: "#/components/schemas/User" }

    # ── Devices ───────────────────────────────────────────────────────
    Device:
      type: object
      required: [id, userId, deviceSerial, name, status, createdAt]
      properties:
        id:           { type: integer }
        userId:       { type: integer }
        deviceSerial: { type: string }
        name:         { type: string }
        status:       { type: string, enum: [online, offline, calibrating] }
        wifiNetwork:  { type: ["string", "null"] }
        createdAt:    { type: string, format: date-time }
    DeviceInput:
      type: object
      required: [deviceSerial, name]
      properties:
        deviceSerial: { type: string }
        name:         { type: string }
        wifiNetwork:  { type: string }
    DeviceUpdate:
      type: object
      properties:
        name:        { type: string }
        status:      { type: string, enum: [online, offline, calibrating] }
        wifiNetwork: { type: string }

    # ── Sensor Readings ───────────────────────────────────────────────
    SensorReading:
      type: object
      required: [id, deviceId, gasLevelPercent, pressurePa, gasDetected, createdAt]
      properties:
        id:             { type: integer }
        deviceId:       { type: integer }
        gasLevelPercent: { type: number }
        pressurePa:     { type: number }
        gasDetected:    { type: boolean }
        createdAt:      { type: string, format: date-time }
    SensorReadingInput:
      type: object
      required: [gasLevelPercent, pressurePa, gasDetected]
      properties:
        gasLevelPercent: { type: number }
        pressurePa:      { type: number }
        gasDetected:     { type: boolean }

    # ── Alerts ────────────────────────────────────────────────────────
    Alert:
      type: object
      required: [id, deviceId, userId, type, message, severity, createdAt]
      properties:
        id:         { type: integer }
        deviceId:   { type: integer }
        userId:     { type: integer }
        type:       { type: string, enum: [gas_leak, low_level, battery, offline] }
        message:    { type: string }
        severity:   { type: string, enum: [info, warning, critical] }
        resolvedAt: { type: ["string", "null"] }
        createdAt:  { type: string, format: date-time }

    # ── Refill Orders ─────────────────────────────────────────────────
...
    Message:
      type: object
      required: [id, senderId, receiverId, content, sentAt]
      properties:
        id:         { type: integer }
        senderId:   { type: integer }
        receiverId: { type: integer }
        content:    { type: string }
        sentAt:     { type: string, format: date-time }
        senderName: { type: string }
    MessageInput:
      type: object
      required: [receiverId, content]
      properties:
        receiverId: { type: integer }
        content:    { type: string }
...