# lib/api-zod/src/generated/types/userRegistration.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Gas Monitor API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRegistrationRole } from './userRegistrationRole';

export interface UserRegistration {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: UserRegistrationRole;
}